Changes for protobuf vehicle locations, remove vehicle stop times
Previously, we fetched vehicle locations one route at a time with the nextbus endpoint. With the new Socrata API, we download a protobuf file with all the vehicle locations.

This also removes all the code related to making vehicle stop times. The code was messy
and the task should be done elsewhere.